refactor: test-utils utils.ts to index.ts
