+++ packages/test-utils/package.json
@@ -2,8 +2,8 @@
   "name": "@streamr/test-utils",
   "version": "7.1.0",
   "description": "a collection of Node.js utilities for dealing with asynchronous behavior in jest/mocha tests",
-  "main": "./dist/src/utils.js",
-  "types": "./dist/src/utils.d.ts",
+  "main": "./dist/src/index.js",
+  "types": "./dist/src/index.d.ts",
   "scripts": {
     "check": "tsc -p ./tsconfig.jest.json --noEmit",
     "clean": "jest --clearCache || true; rm -rf dist *.tsbuildinfo node_modules/.cache || true",
